fixprop type warnings in console, allow for decimal values

+++ src/components/KegAdd.js
@@ -9,9 +9,9 @@ function KegAdd(props) {
     props.onKegAddCreation({
       kegName: event.target.kegName.value,
       kegBrand: event.target.kegBrand.value,
-      kegPrice: parseInt(event.target.kegPrice.value),
-      kegABV: parseInt(event.target.kegABV.value),
-      kegAmt: parseInt(event.target.kegAmt.value),
+      kegPrice: parseFloat(event.target.kegPrice.value),
+      kegABV: parseFloat(event.target.kegABV.value),
+      kegAmt: parseFloat(event.target.kegAmt.value),
       kegId: v4()
     });
   }
